// packages/cli/src/management-api/openapi.d.ts
export interface operations {
// ...
  postV1Projects: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: {
      content: {
        'application/json': {
          /** @default true */
          createDatabase?: boolean
          name?: string
          /** @enum {string} */
          region?: 'us-east-1' | 'us-west-1' | 'eu-west-3' | 'eu-central-1' | 'ap-northeast-1' | 'ap-southeast-1'
        }
      }
    }
    responses: {
      /** @description New project created. */
      201: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/json': {
            data: {
              id: string
              /** @constant */
              type: 'project'
              name: string
              /** Format: date-time */
              createdAt: string
              workspace: {
                id: string
                name: string
              }
              database: {
                id: string
                /** @constant */
                type: 'database'
                name: string
                /** @enum {string} */
                status: 'provisioning' | 'ready'
                /** Format: date-time */
                createdAt: string
                isDefault: boolean
                region: {
                  id: string
                  name: string
                }
                apiKeys: {
                  id: string
                  /** @constant */
                  type: 'connection'
                  name: string
                  /** Format: date-time */
                  createdAt: string
                  connectionString: string
                }[]
                connectionString: string | null
                directConnection: {
                  host: string
                  pass: string
                  user: string
                } | null
              } | null
            }
          }
        }
      }
      /** @description Missing or invalid authorization token. */
      401: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/json': {
            error: {
              code: string
              message: string
            }
          }
        }
      }
    }
  }
// ...
}